--- src/assets/css/app.css
@@ -1,6 +1,0 @@
-[v-cloak] {
-  display: none;
-}
-.line-chart .apexcharts-svg {
-  transform: translateY(3px);
-}
