+++ src/components/Clients/Clients.vue
@@ -1,0 +1,13 @@
+<template>
+  <div
+    v-for="client in clientsStore.clients"
+    :key="client.id"
+    class="relative overflow-hidden border-b last:border-b-0 border-gray-100 dark:border-neutral-600 border-solid"
+  >
+    <Client :client="client" />
+  </div>
+</template>
+
+<script setup lang="ts">
+const clientsStore = useClientsStore();
+</script>
