+++ src/eslint.config.mjs
@@ -1,4 +1,10 @@
 import { createConfigForNuxt } from '@nuxt/eslint-config/flat';
 import eslintConfigPrettier from 'eslint-config-prettier';
 
-export default createConfigForNuxt().append(eslintConfigPrettier);
+export default createConfigForNuxt()
+  .append({
+    rules: {
+      'vue/no-multiple-template-root': 'off',
+    },
+  })
+  .append(eslintConfigPrettier);
