+++ src/server/middleware/session.ts
@@ -1,18 +1,13 @@
 export default defineEventHandler(async (event) => {
-  if (event.node.req.url === undefined) {
-    throw createError({
-      statusCode: 400,
-      statusMessage: 'Invalid request',
-    });
-  }
+  const url = getRequestURL(event);
   if (
     !REQUIRES_PASSWORD ||
-    !event.node.req.url.startsWith('/api/') ||
-    event.node.req.url === '/api/session' ||
-    event.node.req.url === '/api/lang' ||
-    event.node.req.url === '/api/release' ||
-    event.node.req.url === '/api/ui-chart-type' ||
-    event.node.req.url === '/api/ui-traffic-stats'
+    !url.pathname.startsWith('/api/') ||
+    url.pathname === '/api/session' ||
+    url.pathname === '/api/lang' ||
+    url.pathname === '/api/release' ||
+    url.pathname === '/api/ui-chart-type' ||
+    url.pathname === '/api/ui-traffic-stats'
   ) {
     return;
   }
@@ -22,7 +17,7 @@ export default defineEventHandler(async (event) => {
   }
 
   const authorization = getHeader(event, 'Authorization');
-  if (event.node.req.url.startsWith('/api/') && authorization) {
+  if (url.pathname.startsWith('/api/') && authorization) {
     if (isPasswordValid(authorization)) {
       return;
     }
