+++ src/server/plugins/shutdown.ts
@@ -1,0 +1,6 @@
+export default defineNitroPlugin((nitroApp) => {
+  nitroApp.hooks.hook('close', () => {
+    console.log('Shutting down');
+    WireGuard.Shutdown();
+  });
+});
